FastApi_app/requirements.txt
fastapi
uvicorn
paho-mqtt
pymongo